cpc_project/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt
--extra-index-url=http://dist.eldarion.com/gondor/
--extra-index-url=http://dist.gondor.io/dev/

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

gondor==1.0b1.post1
nashvegas==0.6.2
psycopg2==2.4
django-taggit==0.9.3
django-filter==0.5.3
django-groups==0.1.dev10
django-avatar==2.0a1
PIL==1.1.7
django-markup==0.3
dialogos==0.1.dev3
django-pagination==1.0.7
django-voting==0.1
django-notification==0.3a1.dev4
